# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: Microservice for getting services
  description: Ruby on Rails + MySQL
  license:
    name: Software Architecture, 2020-I
host: host:3000
basePath: /ridesupport-getservices-ms/resources
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /vehicle_types/{id}/vehicles/:
    get:
      summary: Returns the vehicles given a id of vehicle type
      operationId: vehiclesByType
      parameters:
        - name: id
          in: path
          description: The vehicles which are of vehicle type
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/VehicleType'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
            description: Not Acceptable (Invalid Params)
            schema:
              $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'  
...
definitions:
...
  VehicleType:
    type: object
    properties:
      vehicle_type_id:
        type: integer
      name:
        type: string
...
  ErrorModel:
    type: object
    required:
      - message
      - code
    properties:
      message:
        type: string
      code:
        type: number
      description:
        type: object